Flask_API/test_swagger_integration.py
"""
Test script for Swagger integration in Flask API
Tests the Swagger-enabled endpoints directly in app.py
"""

import requests
import json

def test_swagger_endpoints():
    """Test the Swagger-enabled API endpoints"""
    
    print("\n" + "="*70)
    print("🧪 Testing Flask API with Swagger Integration")
    print("="*70 + "\n")
    
    base_url = "http://localhost:5000"
    
    # Test endpoints
    endpoints = [
        ("API Info", f"{base_url}/api/"),
        ("Health Check", f"{base_url}/api/health"),
        ("All Members", f"{base_url}/api/members/"),
        ("Members with Tasks", f"{base_url}/api/members/with-tasks"),
        ("All Tasks", f"{base_url}/api/tasks/"),
        ("All Task Items", f"{base_url}/api/task-items/"),
    ]
    
    for name, url in endpoints:
        try:
            print(f"🔍 Testing: {name}")
            print(f"   URL: {url}")
            
            response = requests.get(url)
            
            if response.status_code == 200:
                data = response.json()
                print(f"   ✅ Status: {response.status_code}")
                print(f"   📊 Success: {data.get('success', 'N/A')}")
                print(f"   📝 Message: {data.get('message', 'N/A')}")
                
                if 'data' in data:
                    if isinstance(data['data'], list):
                        print(f"   📋 Items: {len(data['data'])}")
                        if data['data']:
                            sample = data['data'][0]
                            print(f"   📋 Sample keys: {list(sample.keys())}")
                    else:
                        print(f"   📋 Data keys: {list(data['data'].keys())}")
                
            else:
                print(f"   ❌ Status: {response.status_code}")
                print(f"   📄 Response: {response.text[:100]}...")
                
        except requests.exceptions.ConnectionError:
            print(f"   ❌ Connection Error: Flask app not running")
            print(f"   💡 Run: python app.py")
        except Exception as e:
            print(f"   ❌ Error: {str(e)}")
        
        print()

def show_swagger_info():
    """Show information about Swagger features"""
    
    print("🎯 SWAGGER FEATURES:")
    print("="*50)
    print("📚 Interactive Documentation: http://localhost:5000/swagger/")
    print("🔍 Try it out buttons for each endpoint")
    print("📋 Request/Response examples with real data")
    print("✅ Parameter validation and type checking")
    print("🧪 Built-in API testing interface")
    print("📖 Automatic documentation generation")
    print()

def show_usage_instructions():
    """Show how to use the Swagger UI"""
    
    print("🎯 HOW TO USE SWAGGER UI:")
    print("="*50)
    print("1. 🌐 Open: http://localhost:5000/swagger/")
    print("2. 📋 Browse available endpoints")
    print("3. 🔍 Click on any endpoint to expand")
    print("4. 🧪 Click 'Try it out' button")
    print("5. 📝 Fill in parameters (if any)")
    print("6. ▶️  Click 'Execute' button")
    print("7. 📊 See the response with real data!")
    print()
    print("💡 Pro Tips:")
    print("   - Use the 'Try it out' feature to test endpoints")
    print("   - Check the request/response examples")
    print("   - Use filters to get specific data")
    print("   - Test the nested data endpoint: /api/members/with-tasks")
    print()

if __name__ == '__main__':
    print("🎯 Flask API Swagger Integration Test")
    print("This demonstrates the Swagger-enabled API integrated into app.py!")
    
    try:
        test_swagger_endpoints()
        show_swagger_info()
        show_usage_instructions()
        
        print("🎉 SWAGGER INTEGRATION WORKING PERFECTLY!")
        print("="*70)
        print("🚀 Next Steps:")
        print("   1. Run: python app.py")
        print("   2. Open http://localhost:5000/swagger/ in your browser")
        print("   3. Explore the interactive documentation")
        print("   4. Test endpoints using the 'Try it out' feature")
        print("   5. Share the Swagger UI with your team!")
        
    except Exception as e:
        print(f"❌ Error: {str(e)}")
        print("Make sure your Flask app is running: python app.py")
