Flask_API/requirements.txt
# Flask and server
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-CORS>=4.0.0
Flask-RESTX>=1.3.0

# Database and ORM
SQLAlchemy>=2.0.0
PyMySQL>=1.0.2
cryptography>=3.4.8

# Configuration and validation
python-dotenv>=1.0.0
marshmallow>=3.20.0
flask-marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0

# Development and testing
pytest>=7.4.0
pytest-flask>=1.2.0
